move "promote user" underneath admin manager
